put interfaces in component dirs
