+++ angular/package.json
@@ -37,12 +37,12 @@
     "@angular/language-service": "~8.2.0",
     "@babel/traverse": "^7.6.0",
     "@compodoc/compodoc": "^1.1.11",
-    "@storybook/addon-a11y": "^5.3.0-beta.14",
-    "@storybook/addon-docs": "^5.3.0-beta.14",
-    "@storybook/addons": "^5.3.0-beta.14",
-    "@storybook/angular": "^5.3.0-beta.14",
+    "@storybook/addon-a11y": "^5.3.0-beta.16",
+    "@storybook/addon-docs": "^5.3.0-beta.16",
+    "@storybook/addons": "^5.3.0-beta.16",
+    "@storybook/angular": "^5.3.0-beta.16",
     "@storybook/preset-scss": "^1.0.2",
-    "@storybook/theming": "^5.3.0-beta.14",
+    "@storybook/theming": "^5.3.0-beta.16",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
