--- angular/projects/spark-angular/src/lib/interfaces/sprk-footer-award.interface.ts
@@ -1,8 +1,0 @@
-import { TSprkFooterLinkOptionalProps } from './sprk-footer-link.interface';
-import { ISprkFooterImg } from './sprk-footer-img.interface';
-
-/**
- * An image that is wrapped with a link.
- */
-export interface ISprkFooterAward extends TSprkFooterLinkOptionalProps, ISprkFooterImg { }
-
