--- angular/projects/spark-angular/src/lib/interfaces/sprk-footer-img.interface.ts
@@ -1,11 +1,0 @@
-import { ISprkImg } from './sprk-img.interface';
-
-export interface ISprkFooterImg extends ISprkImg {
-  /**
-   * Expects a space separated string
-   * of classes to be added to the
-   * image if `imgSrc` has a value.
-   */
-  imgCSS?: string;
-}
-
