--- angular/projects/spark-angular/src/lib/interfaces/sprk-footer-link.interface.ts
@@ -1,15 +1,0 @@
-import { ISprkLink } from './sprk-link.interface';
-import { TSprkOptionalProps } from './utilities.interface';
-
-export interface ISprkFooterLink extends ISprkLink {
-  /**
-   * TODO: Deprecate this.
-   * The Footer API supports using this
-   * or `analyticsString` but
-   * using `analyticsString` is
-   * preferred.
-   */
-  analytics?: string;
-}
-
-export type TSprkFooterLinkOptionalProps = TSprkOptionalProps<ISprkFooterLink>;
