--- angular/projects/spark-angular/src/lib/interfaces/sprk-icon-interface.ts
@@ -1,6 +1,0 @@
-export interface ISprkIcon {
-  /**
-   * Name of the icon.
-   */
-  icon: string;
-}
