--- angular/projects/spark-angular/src/lib/interfaces/sprk-img.interface.ts
@@ -1,10 +1,0 @@
-export interface ISprkImg {
-  /**
-   * Value for the `src` attribute of the image.
-   */
-  imgSrc: string;
-  /**
-   * Value for the `alt` attribute of the image.
-   */
-  imgAlt: string;
-}
