--- angular/projects/spark-angular/src/lib/interfaces/sprk-narrow-nav-link.interface.ts
@@ -1,18 +1,0 @@
-import { TSprkIconTextLinkOptionalProps } from './sprk-icon-text-link.interface';
-
-/**
- * Structure of a link in
- * the Narrow version of the Masthead.
- * Narrow Nav Links can have an
- * optional Leading Icon.
- */
-export interface ISprkNarrowNavLink extends TSprkIconTextLinkOptionalProps {
-  /**
-   * If `true`, active styles will be applied.
-   */
-  active?: boolean;
-  /**
-   * Optional sub-navigation for the link.
-   */
-  subNav?: TSprkIconTextLinkOptionalProps[];
-}
