--- angular/projects/spark-angular/src/lib/interfaces/sprk-text.interface.ts
@@ -1,10 +1,0 @@
-/**
- * Basic object that contains
- * a key of `text`. Commonly
- * used to build up paragraphs of
- * text.
- * TODO:// Refactor Footer to use something else.
- */
-export interface ISprkText {
-  text: string;
-}
