--- angular/projects/spark-angular/src/lib/interfaces/utilities.interface.ts
@@ -1,6 +1,0 @@
-/**
- * Use to make all properties optional.
- */
-export type TSprkOptionalProps<T> = {
-    [P in keyof T]?: T[P];
-};
